chore: updated

+++ client/src/chunker.rs
@@ -177,9 +177,9 @@ impl Chunker {
 pub struct BytesWeighter;
 
 impl Weighter<String, Vec<u8>> for BytesWeighter {
-    fn weight(&self, _key: &String, val: &Vec<u8>) -> u32 {
+    fn weight(&self, _key: &String, val: &Vec<u8>) -> u64 {
         // Be cautions out about zero weights!
-        val.len().clamp(1, u32::MAX as usize) as u32
+        val.len().clamp(1, u64::MAX as usize) as u64
     }
 }
